#642 Fix a similar issue

+++ Assets/Scripts/PlayerController/ElectricBeamAbility.cs
@@ -294,7 +294,7 @@ public class ElectricBeamAbility : PlayerAbility
             base.CurrentState = value;
             Activated = value.Bool("activated");
             TargetId = value.Int("targetId");
-            Charge = value.Int("charge");
+            Charge = value.Float("charge");
         }
     }
 }
